services/chatbot/requirements.txt
flask
flask-cors
psycopg2-binary
minio
python-dotenv
requests
